1. 오디오 입력 -> AudioListener 패키지로 2. 오디오 출력 -> AudioSpeaker 패키지로로 3. LLM 입력및 응답 -> LLMInterviewer 패키지로 4. main.py 파일에서 전체 어플리케이션을 실행
