--- STT/requirements.txt
@@ -1,4 +1,0 @@
-# 소리를 텍스트 파일로 
-whisper
-scipy
-sounddevice
