--- STT/show_device.py
@@ -1,2 +1,0 @@
-import sounddevice as sd
-print(sd.query_devices())
