--- TTS/requirements.txt
@@ -1,5 +1,0 @@
-# 텍스트를 소리파일로 바꿔주는 것
-gTTS
-
-# 소리파일을 컴퓨터 스피커로 내보내는 것
-playsound
